interactive-knowledge-base/requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-core
langgraph
openai
chromadb
python-dotenv
llama-index
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
psycopg2-binary
pandas
browser-use
